core/task_clustering.py
"""
Neuro-Symbolic Autonomy Framework (NSAF) - Task Clustering
==========================================================

Author: Bolorerdene Bundgaa
Contact: bolor@ariunbolor.org
Website: https://bolor.me

Implements quantum-enhanced symbolic task clustering for decomposing complex problems
into manageable task clusters.
"""

from typing import List, Dict, Any, Optional
import numpy as np
from qiskit import QuantumCircuit, QuantumRegister, ClassicalRegister
from qiskit_aer import AerSimulator
from scipy.spatial.distance import cosine
import networkx as nx
from dataclasses import dataclass
import yaml
import logging
from pathlib import Path

@dataclass
class TaskCluster:
    """Represents a cluster of related tasks"""
    id: str
    tasks: List[Dict[str, Any]]
    centroid: np.ndarray
    similarity_score: float
    quantum_state: Optional[np.ndarray] = None

class QuantumSymbolicTaskClustering:
    """
    Implements quantum-enhanced symbolic task clustering for decomposing complex problems
    into manageable task clusters.
    """
    
    def __init__(self, config_path: str = "config/config.yaml"):
        # Load configuration
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)['task_clustering']
            
        self.max_cluster_size = self.config['max_cluster_size']
        self.min_cluster_size = self.config['min_cluster_size']
        self.similarity_threshold = self.config['similarity_threshold']
        
        # Quantum backend setup
        self.backend = AerSimulator()
        self.shots = self.config['quantum_optimization']['shots']
        
        # Initialize graph for symbolic relationships
        self.task_graph = nx.Graph()
        
        # Setup logging
        logging.basicConfig(level=logging.INFO)
        self.logger = logging.getLogger(__name__)

    def _compute_task_embedding(self, task: Dict[str, Any]) -> np.ndarray:
        """
        Compute the embedding for a task using foundation models.
        
        Args:
            task: Dictionary containing task description and metadata
            
        Returns:
            numpy.ndarray: Task embedding vector
        """
        # TODO: Implement foundation model embedding
        # This is a placeholder that returns random embeddings
        return np.random.randn(768)  # Using 768 as typical embedding dimension

    def _quantum_similarity(self, vec1: np.ndarray, vec2: np.ndarray) -> float:
        """
        Compute similarity between two vectors using quantum circuit.
        
        Args:
            vec1, vec2: Input vectors to compare
            
        Returns:
            float: Quantum-enhanced similarity score
        """
        # Normalize vectors
        vec1_norm = vec1 / np.linalg.norm(vec1)
        vec2_norm = vec2 / np.linalg.norm(vec2)
        
        # Create quantum circuit
        qr = QuantumRegister(2, 'q')
        cr = ClassicalRegister(1, 'c')
        circuit = QuantumCircuit(qr, cr)
        
        # Encode vectors into quantum states
        # This is a simplified encoding - could be enhanced with more sophisticated methods
        theta1 = np.arccos(vec1_norm[0])
        theta2 = np.arccos(vec2_norm[0])
        
        circuit.ry(2 * theta1, qr[0])
        circuit.ry(2 * theta2, qr[1])
        
        # Add interference
        circuit.cx(qr[0], qr[1])
        circuit.h(qr[0])
        
        # Measure
        circuit.measure(qr[0], cr[0])
        
        # Execute circuit
        job = self.backend.run(circuit, shots=self.shots)
        result = job.result()
        counts = result.get_counts(circuit)
        
        # Compute similarity from measurement statistics
        similarity = counts.get('0', 0) / self.shots
        
        return similarity

    def _build_symbolic_graph(self, tasks: List[Dict[str, Any]]) -> None:
        """
        Build a graph representing symbolic relationships between tasks.
        
        Args:
            tasks: List of task dictionaries
        """
        # Clear existing graph
        self.task_graph.clear()
        
        # Add nodes
        for i, task in enumerate(tasks):
            self.task_graph.add_node(i, task=task)
        
        # Add edges based on symbolic relationships
        for i in range(len(tasks)):
            for j in range(i + 1, len(tasks)):
                task1_embedding = self._compute_task_embedding(tasks[i])
                task2_embedding = self._compute_task_embedding(tasks[j])
                
                # Compute hybrid similarity
                classical_sim = 1 - cosine(task1_embedding, task2_embedding)
                quantum_sim = self._quantum_similarity(task1_embedding, task2_embedding)
                
                # Combine classical and quantum similarities
                hybrid_sim = 0.5 * (classical_sim + quantum_sim)
                
                if hybrid_sim >= self.similarity_threshold:
                    self.task_graph.add_edge(i, j, weight=hybrid_sim)

    def _optimize_clusters(self, clusters: List[TaskCluster]) -> List[TaskCluster]:
        """
        Optimize cluster assignments using quantum optimization.
        
        Args:
            clusters: List of initial task clusters
            
        Returns:
            List[TaskCluster]: Optimized task clusters
        """
        # TODO: Implement quantum optimization for cluster refinement
        # This is a placeholder that returns the input clusters
        return clusters

    def decompose_problem(self, problem: Dict[str, Any]) -> List[TaskCluster]:
        """
        Decompose a complex problem into quantum-symbolic task clusters.
        
        Args:
            problem: Dictionary containing problem description and requirements
            
        Returns:
            List[TaskCluster]: List of task clusters
        """
        # Extract tasks from problem
        tasks = self._extract_tasks(problem)
        
        # Build symbolic relationship graph
        self._build_symbolic_graph(tasks)
        
        # Initial clustering using spectral clustering
        n_clusters = max(
            min(len(tasks) // self.min_cluster_size,
                len(tasks) // self.max_cluster_size + 1),
            1
        )
        
        # Use spectral clustering for initial grouping
        clusters = []
        if len(tasks) > 0:
            # Convert graph to adjacency matrix
            adj_matrix = nx.adjacency_matrix(self.task_graph).todense()
            
            # Perform spectral clustering
            from sklearn.cluster import SpectralClustering
            clustering = SpectralClustering(
                n_clusters=n_clusters,
                affinity='precomputed',
                random_state=42
            ).fit(adj_matrix)
            
            # Create task clusters
            for i in range(n_clusters):
                cluster_tasks = [tasks[j] for j in range(len(tasks)) 
                               if clustering.labels_[j] == i]
                if cluster_tasks:
                    # Compute cluster centroid
                    embeddings = [self._compute_task_embedding(task) 
                                for task in cluster_tasks]
                    centroid = np.mean(embeddings, axis=0)
                    
                    # Create cluster
                    cluster = TaskCluster(
                        id=f"cluster_{i}",
                        tasks=cluster_tasks,
                        centroid=centroid,
                        similarity_score=np.mean([
                            1 - cosine(emb, centroid) for emb in embeddings
                        ])
                    )
                    clusters.append(cluster)
        
        # Optimize clusters using quantum optimization
        optimized_clusters = self._optimize_clusters(clusters)
        
        return optimized_clusters

    def _extract_tasks(self, problem: Dict[str, Any]) -> List[Dict[str, Any]]:
        """
        Extract individual tasks from a complex problem.
        
        Args:
            problem: Dictionary containing problem description and requirements
            
        Returns:
            List[Dict[str, Any]]: List of individual tasks
        """
        # TODO: Implement task extraction logic
        # This is a placeholder that assumes the problem contains a 'tasks' key
        return problem.get('tasks', [])

    def visualize_clusters(self, clusters: List[TaskCluster]) -> None:
        """
        Visualize task clusters using networkx.
        
        Args:
            clusters: List of task clusters to visualize
        """
        import matplotlib.pyplot as plt
        
        plt.figure(figsize=(12, 8))
        pos = nx.spring_layout(self.task_graph)
        
        # Draw nodes
        nx.draw_networkx_nodes(self.task_graph, pos)
        
        # Draw edges with weights as colors
        edges = self.task_graph.edges(data=True)
        weights = [e[2]['weight'] for e in edges]
        nx.draw_networkx_edges(self.task_graph, pos, edge_color=weights, 
                             edge_cmap=plt.cm.viridis)
        
        plt.title("Task Cluster Visualization")
        plt.colorbar(plt.cm.ScalarMappable(cmap=plt.cm.viridis), 
                    label="Similarity Score")
        plt.show()
