requirements.txt
# Neuro-Symbolic Autonomy Framework (NSAF) - Unified Requirements
# ================================================================

# Core Scientific Computing
numpy>=1.24.0
scipy>=1.7.0
sympy>=1.8
networkx>=2.6.0

# Machine Learning & Deep Learning
torch>=1.9.0
tensorflow>=2.15.0
scikit-learn>=1.4.0

# Quantum Computing
qiskit>=1.0.1
cirq>=0.13.0
pennylane>=0.21.0

# Distributed Computing
ray>=2.0.0

# Semantic Web & Knowledge Graphs
rdflib>=6.0.0

# Web Framework & API
fastapi==0.109.1
uvicorn==0.27.1
websockets==12.0
pydantic==2.6.1
pydantic[email]==2.6.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# HTTP Clients
aiohttp==3.9.3
httpx==0.27.0
requests>=2.28.0

# Data Processing
pandas>=2.2.0
matplotlib>=3.7.0
tqdm>=4.65.0

# Configuration & Environment
pyyaml>=6.0.0
python-dotenv>=1.0.1

# Optimization
pyomo>=6.1.0
optuna>=2.10.0

# Testing & Development
pytest>=8.0.0
pytest-asyncio>=0.23.5
black>=21.9b0
isort>=5.9.3
mypy>=0.910
flake8>=3.9.2

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.15.2

# Additional Utilities
jinja2==3.1.3
markdown==3.5

# MCP Protocol Support
# (Using simplified implementation, no additional deps needed)